refactor: Extract InnerListItemsSearchAsync to separate service
